+++ src/data/arbitrary/types/mod.rs
@@ -1,0 +1,6 @@
+mod amino_acids;
+mod cigar;
+mod nucleotides;
+
+pub use cigar::*;
+pub use nucleotides::*;