+++ CHANGELOG.md
@@ -10,7 +10,9 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 
 - Added banded Smith Waterman implementation as a more memory-efficient alternative for long sequence alignment
 - Added `HeaderReadable` getter trait for record types
-- Implemented `Arbitrary` for `AlignmentStates` and added additional wrapper type (behind `fuzzing` feature gate)
+- Added `slice_to_ref_range` method for `Alignment`, in order to slice/clip an alignment to a desired portion of the reference
+- Added `CheckedCigar` trait for summing the lengths for operations consuming the query/reference (behind `fuzzing` feature fate)
+- Implemented `Arbitrary` for `AlignmentStates` and `Alignment` along with additional wrapper types (behind `fuzzing` feature gate)
 - Added Viterbi algorithm for `DomainPhmm` (behind `dev-phmm` feature gate)
 - Implemented `Arbitrary` for `SemiLocalPhmm` and `DomainPhmm` (behind `fuzzing` and `dev-phmm` feature gates)
 - Added `sw_scalar_alignment_override` for more flexible testing of dynamic programming alignment algorithms (behind `fuzzing` feature gate)